+++ contrib/opencensus-ext-azure/opencensus/ext/azure/metrics_exporter/standard_metrics/__init__.py
@@ -16,7 +16,7 @@ from opencensus.metrics.export.gauge import Registry
 from opencensus.metrics.export.metric_producer import MetricProducer
 from opencensus.ext.azure.metrics_exporter.standard_metrics.cpu \
     import ProcessorTimeMetric
-from opencensus.ext.azure.metrics_exporter.standard_metrics.dependency \
+from opencensus.ext.azure.metrics_exporter.standard_metrics.http_dependency \
     import DependencyRateMetric
 from opencensus.ext.azure.metrics_exporter.standard_metrics.memory \
     import AvailableMemoryMetric
@@ -24,13 +24,16 @@ from opencensus.ext.azure.metrics_exporter.standard_metrics.process \
     import ProcessCPUMetric
 from opencensus.ext.azure.metrics_exporter.standard_metrics.process \
     import ProcessMemoryMetric
+from opencensus.ext.azure.metrics_exporter.standard_metrics.http_requests \
+    import RequestsRateMetric
 
 # List of standard metrics to track
 STANDARD_METRICS = [AvailableMemoryMetric,
                     DependencyRateMetric,
                     ProcessCPUMetric,
                     ProcessMemoryMetric,
-                    ProcessorTimeMetric]
+                    ProcessorTimeMetric,
+                    RequestsRateMetric]
 
 
 def register_metrics():