+++ contrib/opencensus-ext-azure/tests/test_azure_standard_metrics.py
@@ -15,19 +15,27 @@
 import collections
 import mock
 import requests
+import sys
 import unittest
 
 from opencensus.ext.azure.metrics_exporter import standard_metrics
 from opencensus.trace import execution_context
+if sys.version_info < (3,):
+    from BaseHTTPServer import HTTPServer
+else:
+    from http.server import HTTPServer
 
 ORIGINAL_FUNCTION = requests.Session.request
+ORIGINAL_CONS = HTTPServer.__init__
 
 
 class TestStandardMetrics(unittest.TestCase):
     def setUp(self):
-        standard_metrics.dependency.dependency_map.clear()
+        standard_metrics.http_dependency.dependency_map.clear()
+        standard_metrics.http_requests.requests_map.clear()
         requests.Session.request = ORIGINAL_FUNCTION
-        standard_metrics.dependency.ORIGINAL_REQUEST = ORIGINAL_FUNCTION
+        standard_metrics.http_dependency.ORIGINAL_REQUEST = ORIGINAL_FUNCTION
+        standard_metrics.http_requests.ORIGINAL_CONSTRUCTOR = ORIGINAL_CONS
 
     @mock.patch('opencensus.ext.azure.metrics_exporter'
                 '.standard_metrics.register_metrics')
@@ -40,12 +48,12 @@ class TestStandardMetrics(unittest.TestCase):
         producer = standard_metrics.AzureStandardMetricsProducer()
         metrics = producer.get_metrics()
 
-        self.assertEqual(len(metrics), 5)
+        self.assertEqual(len(metrics), 6)
 
     def test_register_metrics(self):
         registry = standard_metrics.register_metrics()
 
-        self.assertEqual(len(registry.get_metrics()), 5)
+        self.assertEqual(len(registry.get_metrics()), 6)
 
     def test_get_available_memory_metric(self):
         metric = standard_metrics.AvailableMemoryMetric()
@@ -135,21 +143,21 @@ class TestStandardMetrics(unittest.TestCase):
             logger_mock.exception.assert_called()
 
     def test_dependency_patch(self):
-        map = standard_metrics.dependency.dependency_map
-        standard_metrics.dependency.ORIGINAL_REQUEST = lambda x: None
+        map = standard_metrics.http_dependency.dependency_map
+        standard_metrics.http_dependency.ORIGINAL_REQUEST = lambda x: None
         session = requests.Session()
         execution_context.set_is_exporter(False)
-        result = standard_metrics.dependency.dependency_patch(session)
+        result = standard_metrics.http_dependency.dependency_patch(session)
 
         self.assertEqual(map['count'], 1)
         self.assertIsNone(result)
 
     def test_dependency_patch_exporter_thread(self):
-        map = standard_metrics.dependency.dependency_map
-        standard_metrics.dependency.ORIGINAL_REQUEST = lambda x: None
+        map = standard_metrics.http_dependency.dependency_map
+        standard_metrics.http_dependency.ORIGINAL_REQUEST = lambda x: None
         session = mock.Mock()
         execution_context.set_is_exporter(True)
-        result = standard_metrics.dependency.dependency_patch(session)
+        result = standard_metrics.http_dependency.dependency_patch(session)
 
         self.assertIsNone(map.get('count'))
         self.assertIsNone(result)
@@ -167,21 +175,126 @@ class TestStandardMetrics(unittest.TestCase):
         self.assertEqual(rate, 0)
 
     @mock.patch('opencensus.ext.azure.metrics_exporter'
-                '.standard_metrics.dependency.time')
+                '.standard_metrics.http_dependency.time')
     def test_get_dependency_rate(self, time_mock):
         time_mock.time.return_value = 100
-        standard_metrics.dependency.dependency_map['last_time'] = 98
-        standard_metrics.dependency.dependency_map['count'] = 4
+        standard_metrics.http_dependency.dependency_map['last_time'] = 98
+        standard_metrics.http_dependency.dependency_map['count'] = 4
         rate = standard_metrics.DependencyRateMetric.get_value()
 
         self.assertEqual(rate, 2)
 
     @mock.patch('opencensus.ext.azure.metrics_exporter'
-                '.standard_metrics.dependency.time')
+                '.standard_metrics.http_dependency.time')
     def test_get_dependency_rate_error(self, time_mock):
         time_mock.time.return_value = 100
-        standard_metrics.dependency.dependency_map['last_result'] = 5
-        standard_metrics.dependency.dependency_map['last_time'] = 100
+        standard_metrics.http_dependency.dependency_map['last_result'] = 5
+        standard_metrics.http_dependency.dependency_map['last_time'] = 100
         result = standard_metrics.DependencyRateMetric.get_value()
+
+        self.assertEqual(result, 5)
+
+    def test_request_patch(self):
+        map = standard_metrics.http_requests.requests_map
+        func = mock.Mock()
+        new_func = standard_metrics.http_requests.request_patch(func)
+        new_func()
+
+        self.assertEqual(map['count'], 1)
+        self.assertEqual(len(func.call_args_list), 1)
+
+    def test_server_patch(self):
+        standard_metrics. \
+            http_requests. \
+            ORIGINAL_CONSTRUCTOR = lambda x, y, z: None
+        with mock.patch('opencensus.ext.azure.metrics_exporter'
+                        '.standard_metrics.http_requests'
+                        '.request_patch') as request_mock:
+            handler = mock.Mock()
+            handler.do_DELETE.return_value = None
+            handler.do_GET.return_value = None
+            handler.do_HEAD.return_value = None
+            handler.do_OPTIONS.return_value = None
+            handler.do_POST.return_value = None
+            handler.do_PUT.return_value = None
+            result = standard_metrics. \
+                http_requests. \
+                server_patch(None, None, handler)
+            handler.do_DELETE()
+            handler.do_GET()
+            handler.do_HEAD()
+            handler.do_OPTIONS()
+            handler.do_POST()
+            handler.do_PUT()
+
+            self.assertEqual(result, None)
+            self.assertEqual(len(request_mock.call_args_list), 6)
+
+    def test_server_patch_no_methods(self):
+        standard_metrics. \
+            http_requests. \
+            ORIGINAL_CONSTRUCTOR = lambda x, y, z: None
+        with mock.patch('opencensus.ext.azure.metrics_exporter'
+                        '.standard_metrics.http_requests'
+                        '.request_patch') as request_mock:
+            handler = mock.Mock()
+            result = standard_metrics. \
+                http_requests. \
+                server_patch(None, None, handler)
+            handler.do_DELETE()
+            handler.do_GET()
+            handler.do_HEAD()
+            handler.do_OPTIONS()
+            handler.do_POST()
+            handler.do_PUT()
+
+            self.assertEqual(result, None)
+            self.assertEqual(len(request_mock.call_args_list), 0)
+
+    def test_server_patch_no_args(self):
+        standard_metrics \
+            .http_requests \
+            .ORIGINAL_CONSTRUCTOR = lambda x, y: None
+        r = standard_metrics.http_requests.server_patch(None, None)
+
+        self.assertEqual(r, None)
+
+    def test_server_patch_no_handler(self):
+        standard_metrics \
+            .http_requests \
+            .ORIGINAL_CONSTRUCTOR = lambda x, y, z: None
+        r = standard_metrics.http_requests.server_patch(None, None, None)
+
+        self.assertEqual(r, None)
+
+    def test_get_request_rate_metric(self):
+        metric = standard_metrics.RequestsRateMetric()
+        gauge = metric()
+
+        name = '\\ASP.NET Applications(??APP_W3SVC_PROC??)\\Requests/Sec'
+        self.assertEqual(gauge.descriptor.name, name)
+
+    def test_get_request_rate_first_time(self):
+        rate = standard_metrics.RequestsRateMetric.get_value()
+
+        self.assertEqual(rate, 0)
+
+    @mock.patch('opencensus.ext.azure.metrics_exporter'
+                '.standard_metrics.http_requests.time')
+    def test_get_request_rate(self, time_mock):
+        time_mock.time.return_value = 100
+        standard_metrics.http_requests.requests_map['last_time'] = 98
+        standard_metrics.http_requests.requests_map['count'] = 4
+        rate = standard_metrics.RequestsRateMetric.get_value()
+
+        self.assertEqual(rate, 2)
+
+    @mock.patch('opencensus.ext.azure.metrics_exporter'
+                '.standard_metrics.http_requests.time')
+    def test_get_request_rate_error(self, time_mock):
+        time_mock.time.return_value = 100
+        standard_metrics.http_requests.requests_map['last_result'] = 5
+        standard_metrics.http_requests.requests_map['last_time'] = 100
+        result = standard_metrics.RequestsRateMetric.get_value()
 
         self.assertEqual(result, 5)