+++ contrib/opencensus-ext-azure/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Unreleased
+- Standard metrics incoming requests per second
+  ([#758](https://github.com/census-instrumentation/opencensus-python/pull/758))
 
 ## 0.7.0
 Released 2019-07-31